backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aerich[toml]>=0.9.2",
    "apscheduler>=3.11.1",
    "argon2-cffi>=25.1.0",
    "authlib>=1.6.5",
    "catboost>=1.2.8",
    "email-validator>=2.3.0",
    "fastapi>=0.121.1",
    "fastapi-limiter>=0.1.6",
    "fastapi-mail>=1.5.8",
    "pandas>=2.3.3",
    "pydantic>=2.12.4",
    "pyjwt>=1.7.1",
    "python-dotenv>=1.2.1",
    "python-whois>=0.9.6",
    "redis[async]>=7.0.1",
    "scikit-learn==1.6.1",
    "tldextract>=5.3.0",
    "tortoise-orm>=0.25.1",
    "uvicorn>=0.38.0",
]
